moved pojo to proper package
